requirements.txt
Django>=1.8.8,<1.9
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.20
django-celery==3.1.17
django-flatpages-tinymce==0.1.1
django-debug-toolbar==1.4
django-uuidfield==0.5.0
eventbrite==3.3.3
kombu==3.0.34
lxml==3.5.0
Pillow==3.1.1
psycopg2==2.6.1
python-dateutil==2.5.0
python-mimeparse==0.1.4
pytz==2015.7
requests==2.9.1
six==1.10.0
stripe
django-jwt-auth
django-webpack-loader==0.3.3
djangorestframework
markdown
django-filter
graphene-django>=1.0.dev
django-graphiql
django-compressor

# the stripe library requires custom arguments which the requirements.txt file
# parsing apparently doesn't support, so install it manually on the command
# line using:
# pip install --index-url https://code.stripe.com --upgrade stripe
